rename app file using kebab case
